+++ NEWS.md
@@ -1,5 +1,6 @@
 # Unreleased version
 
+- Added a `launch()` function that launches a shiny app for SSVS
 - Add a boolean `color` parameter to `plot()`
 - Don't automatically print the result of `summary()` to the console since it doesn't allow you to properly save the result to a variable